Fix chicken goto coop
