+++ common/src/main/java/net/satisfy/farm_and_charm/core/entity/ai/ChickenGotoAndEnterCoopGoal.java
@@ -9,35 +9,36 @@ import net.satisfy.farm_and_charm.core.block.ChickenCoopBlock;
 import net.satisfy.farm_and_charm.core.block.entity.ChickenCoopBlockEntity;
 import net.satisfy.farm_and_charm.core.entity.ChickenCoopAccess;
 import net.satisfy.farm_and_charm.core.mixin.ChickenAccessor;
+import net.satisfy.farm_and_charm.core.registry.ObjectRegistry;
 
 import java.util.EnumSet;
+import java.util.Random;
 import java.util.Timer;
 import java.util.TimerTask;
 
-public class ChickenEnterCoopGoal extends Goal {
+public class ChickenGotoAndEnterCoopGoal extends Goal {
     private final Chicken chicken;
+    public int eggTime = new Random().nextInt(6000) + 6000;
+    private int tickCounter;
 
-    public ChickenEnterCoopGoal(Chicken chicken) {
+    public ChickenGotoAndEnterCoopGoal(Chicken chicken) {
         this.chicken = chicken;
         this.setFlags(EnumSet.of(Flag.MOVE));
     }
 
     @Override
     public boolean canUse() {
-        //if (!ChickenLayEggUtil.willLayEgg(chicken)) return false;
         if (!((ChickenCoopAccess) chicken).farmAndCharm$hasCoopTarget()) return false;
         BlockPos target = ((ChickenCoopAccess) chicken).farmAndCharm$getCoopTarget();
 
         if (!(chicken.level().getBlockEntity(target) instanceof ChickenCoopBlockEntity coop)) return false;
-        Vec3 chickenPos = chicken.position();
         Vec3 targetCenter = new Vec3(target.getX() + 0.5, target.getY() + 0.5, target.getZ() + 0.5);
+        Vec3 chickenPos = chicken.position();
         double distance = chickenPos.distanceTo(targetCenter);
+        boolean result = coop.hasSpaceForChicken()
+                && !coop.containsChicken(chicken) && --eggTime <= 0;
 
-        boolean result = distance < 1.5
-                && coop.hasSpaceForChicken()
-                && !coop.containsChicken(chicken);
-
-        if (result) {
+        if (result && distance <= 1.5) {
             chicken.setInvisible(true);
             chicken.setNoAi(true);
             System.out.println("ChickenEnterCoopGoal triggered for " + chicken.getUUID());
@@ -71,7 +72,12 @@ public class ChickenEnterCoopGoal extends Goal {
 
     @Override
     public void start() {
+        tickCounter = 0;
+        Vec3 chickenPos = chicken.position();
         BlockPos target = ((ChickenCoopAccess) chicken).farmAndCharm$getCoopTarget();
+        Vec3 targetCenter = new Vec3(target.getX() + 0.5, target.getY() + 0.5, target.getZ() + 0.5);
+        double distance = chickenPos.distanceTo(targetCenter);
+        if (distance >= 1.5) return;
         if (target != null && chicken.level().getBlockEntity(target) instanceof ChickenCoopBlockEntity coop && coop.hasSpaceForChicken()) {
             chicken.level().playSound(null, chicken.blockPosition(), SoundEvents.BEEHIVE_ENTER, chicken.getSoundSource(), 1.0F, 1.0F);
             coop.addChicken(chicken);
@@ -80,7 +86,30 @@ public class ChickenEnterCoopGoal extends Goal {
     }
 
     @Override
-    public boolean canContinueToUse() {
-        return false;
+    public void stop() {
+        tickCounter = 0;
+        chicken.getNavigation().stop();
+    }
+
+    @Override
+    public void tick() {
+        tickCounter++;
+        --eggTime;
+
+        BlockPos coop = ((ChickenCoopAccess) chicken).farmAndCharm$getCoopTarget();
+        if (coop == null) return;
+
+        if (!chicken.level().getBlockState(coop).is(ObjectRegistry.CHICKEN_COOP.get())) {
+            ((ChickenCoopAccess) chicken).farmAndCharm$clearCoopTarget();
+            chicken.getNavigation().stop();
+            return;
+        }
+
+        boolean success = chicken.getNavigation().moveTo(coop.getX(), coop.getY(), coop.getZ(), 1.0);
+
+        if (!success || tickCounter > 1200) {
+            ((ChickenCoopAccess) chicken).farmAndCharm$clearCoopTarget();
+            chicken.getNavigation().stop();
+        }
     }
 }
